+++ tests/testthat/test-plot_map.R
@@ -10,17 +10,17 @@ start_db()
 # enter data -------------------------------------------------------------------
 dbWriteTable(globaltrends_db, "data_score", example_score, append = TRUE)
 
-# plot_bar.exp_score ------------------------------------------------------------
-test_that("plot_bar.exp_score1", {
+# plot_map.exp_score ------------------------------------------------------------
+test_that("plot_map.exp_score1", {
   keywords <- unique(example_score$keyword)[1]
   data <- map_dfr(keywords, export_score)
-  out1 <- plot_bar(data, type = "obs")
+  out1 <- plot_map(data, type = "obs")
   expect_s3_class(out1, "ggplot")
 
   keywords <- unique(example_score$keyword)[1:2]
   data <- map_dfr(keywords, export_score)
   expect_warning(
-    out2 <- plot_bar(data, type = "obs"),
+    out2 <- plot_map(data, type = "obs"),
     "The plot function is limited to 1 keyword\\.\nYou use 2 keywords\\.\nOnly 'amazon' is used\\."
   )
   expect_s3_class(out2, "ggplot")
@@ -28,16 +28,16 @@ test_that("plot_bar.exp_score1", {
   expect_identical(out1$labels, out2$labels)
 })
 
-test_that("plot_bar.exp_score2", {
+test_that("plot_map.exp_score2", {
   keywords <- unique(example_score$keyword)[1]
   data <- map_dfr(keywords, export_score)
-  out1 <- plot_bar(data, type = "sad")
+  out1 <- plot_map(data, type = "sad")
   expect_s3_class(out1, "ggplot")
 
   keywords <- unique(example_score$keyword)[1:2]
   data <- map_dfr(keywords, export_score)
   expect_warning(
-    out2 <- plot_bar(data, type = "sad"),
+    out2 <- plot_map(data, type = "sad"),
     "The plot function is limited to 1 keyword\\.\nYou use 2 keywords\\.\nOnly 'amazon' is used\\."
   )
   expect_s3_class(out2, "ggplot")
@@ -45,16 +45,16 @@ test_that("plot_bar.exp_score2", {
   expect_identical(out1$labels, out2$labels)
 })
 
-test_that("plot_bar.exp_score3", {
+test_that("plot_map.exp_score3", {
   keywords <- unique(example_score$keyword)[1]
   data <- map_dfr(keywords, export_score)
-  out1 <- plot_bar(data, type = "trd")
+  out1 <- plot_map(data, type = "trd")
   expect_s3_class(out1, "ggplot")
 
   keywords <- unique(example_score$keyword)[1:2]
   data <- map_dfr(keywords, export_score)
   expect_warning(
-    out2 <- plot_bar(data, type = "trd"),
+    out2 <- plot_map(data, type = "trd"),
     "The plot function is limited to 1 keyword\\.\nYou use 2 keywords\\.\nOnly 'amazon' is used\\."
   )
   expect_s3_class(out2, "ggplot")
@@ -62,23 +62,22 @@ test_that("plot_bar.exp_score3", {
   expect_identical(out1$labels, out2$labels)
 })
 
-test_that("plot_bar.exp_score4", {
+test_that("plot_map.exp_score4", {
   keywords <- unique(example_score$keyword)[1]
-  data <- map_dfr(keywords, export_score) %>%
-    filter(location == "US")
-  out1 <- plot_bar(data)
-  out2 <- plot_score_bar(data)
+  data <- map_dfr(keywords, export_score)
+  out1 <- plot_map(data)
+  out2 <- plot_score_map(data)
   expect_identical(out1$labels, out2$labels)
 })
 
-# plot_bar.exp_score defaults ---------------------------------------------------
-test_that("plot_bar.exp_score5", {
+# plot_map.exp_score defaults ---------------------------------------------------
+test_that("plot_map.exp_score5", {
   keywords <- unique(example_score$keyword)[1]
   data <- map_dfr(keywords, export_score)
-  out1 <- plot_bar(data)
-  out2 <- plot_bar(data, type = "obs")
-  out3 <- plot_bar(data, type = "sad")
-  out4 <- plot_bar(data, type = "trd")
+  out1 <- plot_map(data)
+  out2 <- plot_map(data, type = "obs")
+  out3 <- plot_map(data, type = "sad")
+  out4 <- plot_map(data, type = "trd")
 
   expect_identical(out1$labels, out2$labels)
   expect_false(identical(out2, out3))
@@ -86,43 +85,43 @@ test_that("plot_bar.exp_score5", {
   expect_false(identical(out3, out4))
 })
 
-# plot_bar.exp_score signals ----------------------------------------------------
-test_that("plot_bar.exp_scoreS1", {
+# plot_map.exp_score signals ----------------------------------------------------
+test_that("plot_map.exp_scoreS1", {
   data <- export_score(keyword = "fc barcelona")
   expect_error(
-    plot_bar(data, type = 1),
+    plot_map(data, type = 1),
     "'type' must be object of type character.\nYou provided an object of type double."
   )
   expect_error(
-    plot_bar(data, type = "A"),
+    plot_map(data, type = "A"),
     "'type' must be either 'obs', 'sad', or 'trd'.\nYou provided A."
   )
   expect_error(
-    plot_bar(data, type = TRUE),
+    plot_map(data, type = TRUE),
     "'type' must be object of type character.\nYou provided an object of type logical."
   )
   expect_error(
-    plot_bar(data, type = sum),
+    plot_map(data, type = sum),
     "'type' must be object of type character.\nYou provided an object of type builtin."
   )
   expect_error(
-    plot_bar(data, type = c("obs", "sad", "trd")),
+    plot_map(data, type = c("obs", "sad", "trd")),
     "'type' must be object of length 1.\nYou provided an object of length 3."
   )
 })
 
-# plot_bar.abnorm_score ---------------------------------------------------------
-test_that("plot_bar.abnorm_score1", {
+# plot_map.abnorm_score ---------------------------------------------------------
+test_that("plot_map.abnorm_score1", {
   data <- export_score(keyword = unique(example_score$keyword)[[1]], location = "US")
   data <- get_abnorm_hist(data)
-  out1 <- plot_bar(data)
+  out1 <- plot_map(data)
   expect_s3_class(out1, "ggplot")
 
   keywords <- unique(example_score$keyword)[1:2]
   data <- map_dfr(keywords, export_score, location = "US")
   data <- get_abnorm_hist(data)
   expect_warning(
-    out2 <- plot_bar(data),
+    out2 <- plot_map(data),
     "The plot function is limited to 1 keyword\\.\nYou use 2 keywords\\.\nOnly 'amazon' is used\\."
   )
   expect_s3_class(out2, "ggplot")
@@ -130,24 +129,24 @@ test_that("plot_bar.abnorm_score1", {
   expect_identical(out1$labels, out2$labels)
 })
 
-test_that("plot_bar.abnorm_score2", {
+test_that("plot_map.abnorm_score2", {
   data <- export_score(keyword = unique(example_score$keyword)[[1]], location = "US")
   data <- get_abnorm_hist(data)
-  out1 <- plot_bar(data)
-  out2 <- plot_abnorm_score_bar(data)
+  out1 <- plot_map(data)
+  out2 <- plot_abnorm_score_map(data)
   expect_identical(out1$labels, out2$labels)
 })
 
-# plot_bar methods --------------------------------------------------------------
-test_that("plot_bar.methods", {
+# plot_map methods --------------------------------------------------------------
+test_that("plot_map.methods", {
   data <- export_control(control = 1)
   expect_error(
-    plot_bar(data),
+    plot_map(data),
     "no applicable method"
   )
   data <- export_object(object = 1)
   expect_error(
-    plot_bar(data),
+    plot_map(data),
     "no applicable method"
   )
 })